Delete confirmation

+++ WebApplicationRazorLibrary/WebApplicationRazorLibrary/Pages/Books/Delete.cshtml.cs
@@ -14,26 +14,40 @@ namespace WebApplicationRazorLibrary.Pages.Books
     {
         private IBooksRepository m_iBooksStorage;
 
+        //[BindProperty] todo nath , should i use it
+        public Book Book { get; private set; }
         public DeleteModel(IBooksRepository iBooksStorage)
         {
             m_iBooksStorage = iBooksStorage;
         }
 
-        /*  
-         * razor page MUST be decalred with @page "{id}"
-         * id is same as Request.RouteValues["id"]
-         * we have changed OnGet return code to IActionResult to handle return
-         */
-        public IActionResult OnGet(long id)
+        // --- todo nath , where does he get id
+        public IActionResult OnPost(long id)
         {
-            Book book = m_iBooksStorage.Delete(id);
-            if(book != null)
+           // long id = Book.Id;
+            if (m_iBooksStorage.Get(id) == null)
             {
-                return RedirectToPage("/Books/Index");
+                return RedirectToPage("/NotFound");
             }
-            
-            
-            return RedirectToPage("/NotFound");
+
+            m_iBooksStorage.Delete(id);
+            return RedirectToPage("/Books/Index");
+        }
+            /*  
+             * razor page MUST be decalred with @page "{id}"
+             * id is same as Request.RouteValues["id"]
+             * we have changed OnGet return code to IActionResult to handle return
+             */
+            public IActionResult OnGet(long id)
+        {
+            Book = m_iBooksStorage.Get(id);
+            if (Book == null)
+            {
+                return RedirectToPage("/NotFound");
+            }
+
+
+            return Page();
         }
     }
 }
